Rename AirTable to DataTable
Also move Airtable specific code into a concern
