+++ app/models/processor.rb
@@ -1,4 +1,4 @@
-class Processor < AirTable
+class Processor < DataTable
   self.air_table_name = "Processors"
 
   has_many :agreement_processors, dependent: :delete_all
