+++ db/migrate/20240723114025_rename_air_tables.rb
@@ -1,0 +1,5 @@
+class RenameAirTables < ActiveRecord::Migration[7.0]
+  def change
+    rename_table(:air_tables, :data_tables)
+  end
+end
