+++ spec/models/data_table_spec.rb
@@ -1,11 +1,11 @@
 require "rails_helper"
 
-RSpec.describe AirTable, type: :model do
+RSpec.describe DataTable, type: :model do
   describe ".populate" do
     # For behaviour in subclasses use shared example "is_air_table" in those class specs.
 
     it "raises an error" do
-      expect { described_class.populate }.to raise_error("Cannot run on root class - AirTable")
+      expect { described_class.populate }.to raise_error("Cannot run on root class - DataTable")
     end
   end
 end
